Move files for Form and ChangePassword
